requirements.txt

# Plaud API & OAuth
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.7.0
sqlalchemy>=2.0.0
pytest>=7.4.0

# AI Processing
# Use a Responses API compatible version of the OpenAI SDK.
openai>=1.63.0
google-generativeai>=0.3.0
google-genai>=1.0.0

# Vector Database
qdrant-client>=1.12.0

# Model Context Protocol server
mcp>=0.1.0

# Visualization
pyvis>=0.3.2
networkx>=3.0
matplotlib>=3.7.0
plotly>=5.18.0
wordcloud>=1.9.0
pillow>=10.0.0

# Streamlit for Chronos UI
streamlit>=1.28.0
streamlit-timeline>=0.0.2

# ─────────────────────────────────────────────────────────────
# Audio Processing (optional - install for audio features)
# ─────────────────────────────────────────────────────────────
# Whisper for transcription with speaker diarization
openai-whisper>=20231117
# CLAP for audio embeddings (audio-to-audio similarity)
laion-clap>=1.1.4
# Audio I/O utilities
soundfile>=0.12.1
librosa>=0.10.1
pydub>=0.25.1
# Note: Whisper requires ffmpeg installed on system
# macOS: brew install ffmpeg
# Ubuntu: apt install ffmpeg

# ─────────────────────────────────────────────────────────────
# Document Processing (optional - install for PDF/vision features)
# ─────────────────────────────────────────────────────────────
# PDF to image conversion (requires poppler system library)
pdf2image>=1.16.3
# Alternative: PyMuPDF (no system dependencies)
PyMuPDF>=1.23.0

llama-index-llms-gemini>=0.6.0
notion-client>=2.7.0
